Alter confusing wording around verify_client
Ngnix is often used to "terminate TLS", which means that it does all the stuff required for the TLS handshake between client and server, then forwards requests to an upstream service that is not using TLS. This common use of the phrase "terminate TLS" conflicts with the wording in the `verify_client` section of the doc: in a situation where client certificate verification fails, Nginx actually DOES "terminate TLS", in that it completes the TLS handshake and establishes a secure session between the client and the server. It then allows the admin to control what happens after that. This pull request removes the word "terminate" to avoid this confusion.